Refactor code, add spinner to map
Refactor code so that functions common to all sheets are in a separate
file (/js/common.js) - note that calls to these functions are still in
each individual page's file.

Add spinner to map page to make it clear data is still loading whilst
waiting for map and/or slopes to appear.

Add 'overflow: hidden' property to the main chart SVG, which prevents
the map from showing outside the box boundary.

TODO: Fix scroll/zoom of map in Windows.